+++ zcash_primitives/src/transaction/components.rs
@@ -1,6 +1,14 @@
 //! Structs representing the components within Zcash transactions.
+#[cfg(feature = "orchard")]
 pub mod orchard;
+/// Parsing and serialization of Orchard components when the `orchard` feature is disabled.
+#[cfg(not(feature = "orchard"))]
+pub mod orchard_raw;
+#[cfg(feature = "sapling")]
 pub mod sapling;
+/// Parsing and serialization of Sapling components when the `sapling` feature is disabled.
+#[cfg(not(feature = "sapling"))]
+pub mod sapling_raw;
 pub mod sprout;
 
 pub use self::sprout::JsDescription;